refactor(cli): extract causal command family
